수정,삭제 권한체크, service에서 권한체크 하도록

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -63,7 +63,7 @@ public class UsrArticleController {
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 하고 써");
+			return ResultData.from("F-A", "로그인을 한 다음 작성할 수 있습니다.");
 		}
 
 		if (Ut.isEmptyOrNull(title)) {
@@ -85,6 +85,7 @@ public class UsrArticleController {
 	} // /usr/article/doWrite?title=제목&body=내용
 	
 	
+	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 삭제
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
 	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
@@ -92,15 +93,17 @@ public class UsrArticleController {
 		boolean isLogined = false;
 		int loginedMemberId = 0;
 
+		// 로그인 체크
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 하고 써");
+			return ResultData.from("F-A", "로그인을 하고 삭제할 수 있습니다.");
 		}
 		
+		// 유무 체크
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
@@ -108,6 +111,12 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 글이 없어 삭제되지 않았습니다.", id), id);
 		}
 
+		// 권한 체크
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 대한 권한이 없습니다", id));
+		}
+		
+		// 삭제
 		articleService.deleteArticle(id);
 
 //		return id + "번 글이 삭제되었습니다.";
@@ -115,6 +124,8 @@ public class UsrArticleController {
 
 	} // /usr/article/doDelete?id=1
 
+	
+	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 수정
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
@@ -122,22 +133,30 @@ public class UsrArticleController {
 		boolean isLogined = false;
 		int loginedMemberId = 0;
 
+		// 로그인 체크
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 하고 써");
+			return ResultData.from("F-A", "로그인을 하고 수정할 수 있습니다.");
 		}
 		
 		// 수정되기 전 게시글
 		Article article = articleService.getArticleById(id);
 
+		// 게시글 유무 체크
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글을 찾을 수 없어 수정되지 않았습니다.", id));
 		}
 
+		// 권한 체크
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 대한 권한이 없습니다", article.getId()));
+		}
+		
+		// 수정 
 		articleService.modifyArticle(id, title, body);
 
 		// 수정된 게시글 다시 불러옴
